### 2025-08-05 #8718 @catalamarti
Thank you for the opportunity to suggest potential reviewers.

This work is an R package designed to work with OMOP Common Data Model (CDM) data, a widely used CDM in epidemiology and in generating Real World Evidence.

Accordingly, I would like to suggest reviewers who have experience working with OMOP and R packages:
- Raivo Kolde, Marek Oja, Markus Haug (University of Tartu): They have worked on R packages related to OMOP. GitHub usernames: raivokolde, marek05, haugmarkus.
- Javier Gracia-Tabuenca (FinGenn): He also works on OMOP and R packages. GitHub username: javier-gracia-tabuenca-tuni.
- Martijn Schuemie, Jenna Reps, Christopher Knoll (Johnson & Johnson): Developers of several R/Java packages for OMOP. GitHub usernames: schuemie, jreps, chrisknoll.
- Louisa Smith (Northeastern University): She is currently reviewing a related submission ([omock](https://github.com/openjournals/joss-reviews/issues/8178)) from our group. While perhaps not a first choice due to potential overlap she would be a nice reviewer. GitHub username: louisahsmith.

Additionally, the JOSS reviewer list includes several researchers with expertise in R and epidemiology who could be considered: https://reviewers.joss.theoj.org/reviewers/search?area_id=&keywords=Epidemiology&language=29&name=

Please let me know if any further details would be helpful!
