### 2025-10-02 3695e18
Update NEWS.md
