+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   demographics: demographics.html
   summarise: summarise.html
   table-intersect: table-intersect.html
-last_built: 2025-07-09T16:17Z
+last_built: 2025-10-02T14:04Z
 urls:
   reference: https://darwin-eu.github.io/PatientProfiles/reference
   article: https://darwin-eu.github.io/PatientProfiles/articles
