### 2025-11-20 #8718 @goldingn
Hi @raivokolde, @marek05, @haugmarkus would one of you be interested in reviewing this R package related to OMOP that has been submitted to JOSS?

If you aren't familiar, JOSS reviews are fun and pretty easy and take place interactively on GitHub via checklists and issues. Here's an [example](https://github.com/openjournals/joss-reviews/issues/8445), and here's the [manual](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
