### 2025-12-19 #9654 @editorialbot
[REVIEW]: PatientProfiles: An R package to identify patient characteristics in data mapped to the OMOP common data model
**Submitting author:** @catalamarti (<a href="http://orcid.org/0000-0003-3308-9905">Martí Català</a>)
**Repository:** https://github.com/darwin-eu/PatientProfiles
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.2
