### 2025-12-19 #8718 @catalamarti
> but currently have no access to an OMOP environment.

@pacoramon you can use omock: https://ohdsi.github.io/omock/reference/mockCdmFromDataset.html to get some OMOP CDM synthetic datasets in duckdb:
```
library(omock)
cdm <- mockCdmFromDataset(dataset = "GiBleed", source = "duckdb")
```
