### 2026-01-08 darwin-eu/PatientProfiles#15 @javier-gracia-tabuenca-tuni
Reviewer 1 from JOSS: comments on paper.md
Review

This manuscript presents the PatientProfiles R package, which provides useful and well-structured functionality for extracting, intersecting, and summarising patient-level characteristics from OMOP CDM–mapped data. The package adds clear value within the DARWIN tools ecosystem by reducing the need for bespoke code for common cohort characterisation tasks and by promoting standardised, reusable workflows.

The manuscript does not include a **Research Impact Statement** or an **AI usage disclosure**. However, these sections were not explicitly required under the 2025 software paper guidelines, and their absence should therefore not be considered a formal shortcoming.

One area that could be strengthened is the **State of the Field**. The paper would benefit from a brief discussion of existing related tools within the OMOP ecosystem, particularly HADES, and attempts to extend HADES limitations, such as https://github.com/FINNGEN/HadesExtras. While HADES has different aims—focusing more on building complete analytical pipelines rather than interactive data exploration—it represents a commonly used alternative and supports a wider range of databases. For example, when testing the tutorial in our BigQuery CDM I found problems https://github.com/darwin-eu/PatientProfiles/issues/16. Explicitly positioning PatientProfiles relative to HADES and similar tools would help clarify the package’s unique scope and intended use cases.

Overall, this is a clearly written and technically sound manuscript describing a useful software contribution, which would be further improved by a short contextual comparison with existing OMOP tooling.
