### 2026-01-08 #9654 @javier-gracia-tabuenca-tuni
Review in https://github.com/darwin-eu/PatientProfiles/issues/15
