### 2026-01-16 #9654 @goldingn
Thanks @javier-gracia-tabuenca-tuni for your review!

@catalamarti, feel free to get started addressing those comments while waiting for the other review.

@pacoramon, don't forget you need to write a message with `@editorialbot generate my checklist` to start your review checklist
