+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: PatientProfiles
 Type: Package
 Title: Identify Characteristics of Patients in the OMOP Common Data Model
-Version: 1.4.3
+Version: 1.4.5
 Authors@R: c(
     person("Martí", "Català", , "marti.catalasabate@ndorms.ox.ac.uk", 
     role = c("aut", "cre"), comment = c(ORCID = "0000-0003-3308-9905")),
@@ -26,7 +26,7 @@ Description: Identify the characteristics of patients in data mapped to the
 License: Apache License (>= 2)
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 Suggests: 
     bit64,
     CDMConnector (>= 1.3.1),
