+++ reference/reexports.md
@@ -1,0 +1,9 @@
+# Objects exported from other packages
+
+These objects are imported from other packages. Follow the links below
+to see their documentation.
+
+- omopgenerics:
+
+  [`settings`](https://darwin-eu.github.io/omopgenerics/reference/settings.html),
+  [`suppress`](https://darwin-eu.github.io/omopgenerics/reference/suppress.html)
