+++ articles/table-intersect.md
@@ -1,0 +1,234 @@
+# Adding table intersections
+
+So far we’ve seen that we can add variables indicating intersections
+based on cohorts or concept sets. One additional option we have is to
+simply add an intersection based on a table.
+
+Let’s again create a cohort containing people with an ankle sprain.
+
+``` r
+library(CodelistGenerator)
+library(PatientProfiles)
+library(dplyr)
+library(CohortConstructor)
+library(ggplot2)
+library(omock)
+
+cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")
+
+cdm$ankle_sprain <- conceptCohort(
+  cdm = cdm,
+  name = "ankle_sprain",
+  conceptSet = list("ankle_sprain" = 81151)
+)
+cdm$ankle_sprain
+#> # Source:   table<results.test_ankle_sprain> [?? x 4]
+#> # Database: DuckDB 1.4.4 [unknown@Linux 6.11.0-1018-azure:R 4.5.2//tmp/RtmpKTJczE/file236614828c4b.duckdb]
+#>    cohort_definition_id subject_id cohort_start_date cohort_end_date
+#>                   <int>      <int> <date>            <date>         
+#>  1                    1       2097 1982-11-04        1982-12-02     
+#>  2                    1       2406 1958-01-20        1958-02-17     
+#>  3                    1       2437 2018-04-04        2018-05-02     
+#>  4                    1       2651 1951-06-24        1951-07-29     
+#>  5                    1       3219 1990-08-24        1990-09-28     
+#>  6                    1       4649 1978-08-22        1978-09-05     
+#>  7                    1       5056 2013-09-30        2013-10-14     
+#>  8                    1       1545 2013-06-25        2013-07-23     
+#>  9                    1       2915 1974-01-12        1974-01-26     
+#> 10                    1       3218 1966-01-09        1966-01-30     
+#> # ℹ more rows
+
+cdm$ankle_sprain |>
+  addTableIntersectFlag(
+    tableName = "condition_occurrence",
+    window = c(-30, -1)
+  ) |>
+  glimpse()
+#> Rows: ??
+#> Columns: 5
+#> Database: DuckDB 1.4.4 [unknown@Linux 6.11.0-1018-azure:R 4.5.2//tmp/RtmpKTJczE/file236614828c4b.duckdb]
+#> $ cohort_definition_id           <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
+#> $ subject_id                     <int> 2097, 2651, 1545, 3486, 412, 3537, 1356…
+#> $ cohort_start_date              <date> 1982-11-04, 1951-06-24, 2013-06-25, 19…
+#> $ cohort_end_date                <date> 1982-12-02, 1951-07-29, 2013-07-23, 19…
+#> $ condition_occurrence_m30_to_m1 <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
+```
+
+We can use table intersection functions to check whether someone had a
+record in the drug exposure table in the 30 days before their ankle
+sprain. If we set targetStartDate to “drug_exposure_start_date” and
[…]
