+++ articles/summarise.md
@@ -1,0 +1,537 @@
+# Summarise result
+
+## Introduction
+
+In previous vignettes we have seen how to add patient level demographics
+(age, sex, prior observation, …) or intersections with cohorts ,
+concepts and tables.
+
+Once we have added several columns to our table of interest we may want
+to summarise all this data into a `summarised_result` object using
+several different estimates.
+
+### Variables type
+
+We support different types of variables, variable type is assigned using
+[`dplyr::type_sum`](https://pillar.r-lib.org/reference/type_sum.html):
+
+- Date: `date` or `dttm`.
+
+- Numeric: `dbl` or `drtn`.
+
+- Integer: `int` or `int64`.
+
+- Categorical: `chr`, `fct` or `ord`.
+
+- Logical: `lgl`.
+
+### Estimates names
+
+We can summarise this data using different estimates:
+
+| Estimate name           | Description                                                                                      | Estimate type |
+|-------------------------|--------------------------------------------------------------------------------------------------|---------------|
+| date                    |                                                                                                  |               |
+| mean                    | mean of the variable of interest.                                                                | date          |
+| sd                      | standard deviation of the variable of interest.                                                  | numeric       |
+| median                  | median of the variable of interest.                                                              | date          |
+| qXX                     | qualtile of XX% the variable of interest.                                                        | date          |
+| min                     | minimum of the variable of interest.                                                             | date          |
+| max                     | maximum of the variable of interest.                                                             | date          |
+| count_missing           | number of missing values.                                                                        | integer       |
+| percentage_missing      | percentage of missing values                                                                     | percentage    |
+| density                 | density distribution                                                                             | multiple      |
+| numeric                 |                                                                                                  |               |
+| sum                     | sum of all the values for the variable of interest.                                              | numeric       |
+| mean                    | mean of the variable of interest.                                                                | numeric       |
+| sd                      | standard deviation of the variable of interest.                                                  | numeric       |
+| median                  | median of the variable of interest.                                                              | numeric       |
+| qXX                     | qualtile of XX% the variable of interest.                                                        | numeric       |
+| min                     | minimum of the variable of interest.                                                             | numeric       |
+| max                     | maximum of the variable of interest.                                                             | numeric       |
+| count_missing           | number of missing values.                                                                        | integer       |
+| percentage_missing      | percentage of missing values                                                                     | percentage    |
+| count                   | count number of \`1\`. Only allowed for binary numeric variables.                                | integer       |
+| percentage              | percentage of occurrences of \`1\` (NA are excluded). Only allowed for binary numeric variables. | percentage    |
[…]
