+++ reference/addAge.md
@@ -79,7 +79,7 @@ cdm <- mockPatientProfiles(source = "duckdb")
 
 cdm$cohort1 |>
   addAge()
-#> # Source:   table<og_001_1771936164> [?? x 5]
+#> # Source:   table<og_001_1771937916> [?? x 5]
 #> # Database: DuckDB 1.4.4 [unknown@Linux 6.11.0-1018-azure:R 4.5.2/:memory:]
 #>    cohort_definition_id subject_id cohort_start_date cohort_end_date   age
 #>                   <int>      <int> <date>            <date>          <int>
