+++ reference/addPriorObservation.md
@@ -52,7 +52,7 @@ cdm <- mockPatientProfiles(source = "duckdb")
 
 cdm$cohort1 |>
   addPriorObservation()
-#> # Source:   table<og_126_1771936258> [?? x 5]
+#> # Source:   table<og_126_1771938014> [?? x 5]
 #> # Database: DuckDB 1.4.4 [unknown@Linux 6.11.0-1018-azure:R 4.5.2/:memory:]
 #>    cohort_definition_id subject_id cohort_start_date cohort_end_date
 #>                   <int>      <int> <date>            <date>         
