+++ man/addBirthday.Rd
@@ -1,32 +1,26 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/addDemographics.R
-\name{addAge}
-\alias{addAge}
-\title{Compute the age of the individuals at a certain date}
+% Please edit documentation in R/addBirthDay.R
+\name{addBirthday}
+\alias{addBirthday}
+\title{Add the birth day of an individual to a table}
 \usage{
-addAge(
+addBirthday(
   x,
-  indexDate = "cohort_start_date",
-  ageName = "age",
-  ageGroup = NULL,
-  ageMissingMonth = 1,
-  ageMissingDay = 1,
+  birthday = 0,
+  birthdayName = "birthday",
+  ageMissingMonth = 1L,
+  ageMissingDay = 1L,
   ageImposeMonth = FALSE,
   ageImposeDay = FALSE,
-  ageUnit = "years",
-  missingAgeGroupValue = "None",
   name = NULL
 )
 }
 \arguments{
 \item{x}{Table with individuals in the cdm.}
 
-\item{indexDate}{Variable in x that contains the date to compute the
-demographics characteristics.}
+\item{birthday}{Number of birth day.}
 
-\item{ageName}{Age variable name.}
-
-\item{ageGroup}{if not NULL, a list of ageGroup vectors.}
+\item{birthdayName}{Birth day variable name.}
 
 \item{ageMissingMonth}{Month of the year assigned to individuals with missing
 month of birth.}
@@ -40,26 +34,31 @@ will be considered as missing for all the individuals.}
 \item{ageImposeDay}{TRUE or FALSE. Whether the day of the date of birth
 will be considered as missing for all the individuals.}
 
-\item{ageUnit}{Unit for age it can either be 'years', 'months' or 'days'.}
-
-\item{missingAgeGroupValue}{Value to include if missing age.}
-
 \item{name}{Name of the new table, if NULL a temporary table is returned.}
 }
 \value{
-tibble with the age column added.
+The table with a new column containing the birth day.
 }
 \description{
-Compute the age of the individuals at a certain date
+\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
+
+The function accounts for leap years and corrects the invalid dates to the
+next valid date.
 }
 \examples{
 \donttest{
 library(PatientProfiles)
+library(dplyr)
 
 cdm <- mockPatientProfiles(source = "duckdb")
 
 cdm$cohort1 |>
-  addAge()
+  addBirthday() |>
+  glimpse()
 
+cdm$cohort1 |>
+  addBirthday(birthday = 5, birthdayName = "bithday_5th") |>
+  glimpse()
 }
+
 }
