+++ reference/addObservationPeriodIdQuery.md
@@ -42,19 +42,19 @@ cdm <- mockPatientProfiles(source = "duckdb")
 cdm$cohort1 |>
   addObservationPeriodIdQuery()
 #> # Source:   SQL [?? x 5]
-#> # Database: DuckDB 1.4.4 [unknown@Linux 6.11.0-1018-azure:R 4.5.2/:memory:]
+#> # Database: DuckDB 1.4.4 [unknown@Linux 6.14.0-1017-azure:R 4.5.2/:memory:]
 #>    cohort_definition_id subject_id cohort_start_date cohort_end_date
 #>                   <int>      <int> <date>            <date>         
-#>  1                    2          3 1968-11-21        1969-01-24     
-#>  2                    1          8 1972-03-31        1979-02-22     
-#>  3                    1          9 1977-03-18        1989-03-10     
-#>  4                    2          7 1918-07-15        1920-02-12     
-#>  5                    3          2 1960-10-12        1964-04-30     
-#>  6                    1          5 1961-02-07        1961-12-05     
-#>  7                    3          4 1932-07-14        1937-01-20     
-#>  8                    1          1 1962-12-31        1967-03-23     
-#>  9                    3          6 1995-05-04        1996-08-30     
-#> 10                    2         10 1909-12-15        1914-12-25     
+#>  1                    1         10 1962-08-08        1963-05-20     
+#>  2                    2          4 1909-10-22        1922-07-03     
+#>  3                    3          7 1998-09-06        2001-08-17     
+#>  4                    2          2 1962-06-29        1975-04-09     
+#>  5                    1          6 1939-10-17        1944-10-16     
+#>  6                    1          5 1956-12-18        1958-11-18     
+#>  7                    2          1 1948-06-19        1950-07-03     
+#>  8                    3          3 1949-06-07        1953-08-21     
+#>  9                    1          9 1939-07-09        1940-05-30     
+#> 10                    3          8 1957-08-22        1957-10-02     
 #> # ℹ 1 more variable: observation_period_id <int>
 
 # }
