+++ reference/addTableIntersectFlag.md
@@ -80,7 +80,7 @@ cdm <- mockPatientProfiles(source = "duckdb")
 
 cdm$cohort1 |>
   addTableIntersectFlag(tableName = "visit_occurrence")
-#> # Source:   table<og_164_1772095391> [?? x 5]
+#> # Source:   table<og_164_1772095786> [?? x 5]
 #> # Database: DuckDB 1.4.4 [unknown@Linux 6.14.0-1017-azure:R 4.5.2/:memory:]
 #>    cohort_definition_id subject_id cohort_start_date cohort_end_date
 #>                   <int>      <int> <date>            <date>         
