+++ articles/table-intersect.md
@@ -23,19 +23,19 @@ cdm$ankle_sprain <- conceptCohort(
 )
 cdm$ankle_sprain
 #> # Source:   table<results.test_ankle_sprain> [?? x 4]
-#> # Database: DuckDB 1.4.4 [unknown@Linux 6.14.0-1017-azure:R 4.5.2//tmp/RtmpFcBILs/file20c15822d4c7.duckdb]
+#> # Database: DuckDB 1.4.4 [unknown@Linux 6.14.0-1017-azure:R 4.5.2//tmp/Rtmp42pxwG/file20bc5f158305.duckdb]
 #>    cohort_definition_id subject_id cohort_start_date cohort_end_date
 #>                   <int>      <int> <date>            <date>         
-#>  1                    1        140 2015-11-09        2015-11-23     
-#>  2                    1        340 2007-10-29        2007-11-26     
-#>  3                    1        412 1958-03-20        1958-04-03     
-#>  4                    1        580 1946-02-07        1946-02-28     
-#>  5                    1       1636 1967-05-13        1967-06-10     
-#>  6                    1       1778 2015-01-14        2015-02-04     
-#>  7                    1       2445 1969-01-28        1969-02-11     
-#>  8                    1       2498 1972-08-03        1972-08-17     
-#>  9                    1       3537 1974-12-05        1975-01-02     
-#> 10                    1       3832 2001-01-26        2001-02-09     
+#>  1                    1       3249 1998-08-22        1998-09-19     
+#>  2                    1       3655 1997-08-04        1997-08-25     
+#>  3                    1       3759 2008-12-21        2009-01-25     
+#>  4                    1       4160 2017-06-08        2017-06-29     
+#>  5                    1       4625 1977-04-17        1977-05-08     
+#>  6                    1       4998 2007-03-28        2007-04-18     
+#>  7                    1        293 2018-03-04        2018-03-25     
+#>  8                    1        311 2016-09-18        2016-10-23     
+#>  9                    1        931 1974-01-04        1974-02-08     
+#> 10                    1       1137 1988-10-09        1988-11-06     
 #> # ℹ more rows
 
 cdm$ankle_sprain |>
@@ -46,11 +46,11 @@ cdm$ankle_sprain |>
   glimpse()
 #> Rows: ??
 #> Columns: 5
-#> Database: DuckDB 1.4.4 [unknown@Linux 6.14.0-1017-azure:R 4.5.2//tmp/RtmpFcBILs/file20c15822d4c7.duckdb]
+#> Database: DuckDB 1.4.4 [unknown@Linux 6.14.0-1017-azure:R 4.5.2//tmp/Rtmp42pxwG/file20bc5f158305.duckdb]
 #> $ cohort_definition_id           <int> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, …
[…]
