### 2026-03-23 #9654 @pacoramon
Dear @goldingn,
Thank you for your patience. I have not been able to get on with this review for a long time. Please, let me apologize for that to the author(s) of the manuscript and the R package.

I have completed the Review checklist without finding any issue worth mentioning during my revision. The only thing I have encountered is an error when trying to implement the examples provided in the manuscript (see regex profile below), regarding the dates of the mock data generated being out of the scope of the dates expected in the Cohort.

I have not found any other error when reproducing the examples provided in the repository, thus I find this may be easily solved by the authors.

Best regards, and sorry again for taking so long to submit my review,
Paco

``` r
pacman::p_install("PatientProfiles")
#> Installing package into 'C:/Users/festupinnan-iacs/AppData/Local/R/win-library/4.5'
#> (as 'lib' is unspecified)
#> Warning: unable to access index for repository http://www.stats.ox.ac.uk/pub/RWin/bin/windows/contrib/4.5:
#>   cannot open URL 'http://www.stats.ox.ac.uk/pub/RWin/bin/windows/contrib/4.5/PACKAGES'
#> package 'PatientProfiles' successfully unpacked and MD5 sums checked
#>
#> The downloaded binary packages are in
#>  C:\Users\festupinnan-iacs\AppData\Local\Temp\RtmpaKBIDs\downloaded_packages
#>
#> PatientProfiles installed

citation("PatientProfiles")
#> To cite package 'PatientProfiles' in publications use:
#>
#>   Català M, Guo Y, Du M, Lopez-Guell K, Burn E, Mercade-Besora N
#>   (2026). _PatientProfiles: Identify Characteristics of Patients in the
#>   OMOP Common Data Model_. doi:10.32614/CRAN.package.PatientProfiles
#>   <https://doi.org/10.32614/CRAN.package.PatientProfiles>, R package
#>   version 1.5.0, <https://CRAN.R-project.org/package=PatientProfiles>.
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Manual{,
#>     title = {PatientProfiles: Identify Characteristics of Patients in the OMOP Common Data
#> Model},
#>     author = {Martí Català and Yuchen Guo and Mike Du and Kim Lopez-Guell and Edward Burn and Nuria Mercade-Besora},
#>     year = {2026},
#>     note = {R package version 1.5.0},
#>     url = {https://CRAN.R-project.org/package=PatientProfiles},
#>     doi = {10.32614/CRAN.package.PatientProfiles},
#>   }

library(PatientProfiles)
#> Warning: package 'PatientProfiles' was built under R version 4.5.3
library(tidyverse)

cdm <- mockPatientProfiles(numberIndividuals = 10000)
#> Warning in validateCdmReference(cdm, soft = .softValidation): There are observation period end dates after the current date: 2026-03-23
#> ℹ The latest max observation period end date found is 2040-06-22
# to customise cohorts

cdm$my_flu_cohort <- cdm$cohort1 |>
    dplyr::filter(cohort_definition_id == 1L) |>
    omopgenerics::newCohortTable(cohortSetRef = dplyr::tibble(
        cohort_definition_id = 1L, cohort_name = "flu"
    ))
#> Error in `validateGeneratedCohortSet()`:
#> ! Present cohort_definition_id must be the same:
#> • cohort_set: 1
#> • cohort_attrition: 1, 2, and 3
```
